Fix some bugs

+++ Metrics/ReturnData/ChapinMetricReturn.cs
@@ -10,7 +10,7 @@ namespace Metrics.ChapinMetric
 		public Dictionary<string, int> VariableCount { get; set; }
 		public Dictionary<string, List<string>> ChapinIOTypes { get; set; }
 		public Dictionary<string, int> VariableIOCount { get; set; }
-		public int MetricResult { get; set; }
-		public int MetricIOResult { get; set; }
+		public double MetricResult { get; set; }
+		public double MetricIOResult { get; set; }
 	}
 }
